extracted common classes
