+++ src/main/java/com/marginallyclever/makelangelo/makeart/turtlegenerator/TurtleGeneratorFactory.java
@@ -1,5 +1,8 @@
 package com.marginallyclever.makelangelo.makeart.turtlegenerator;
 
+import com.marginallyclever.makelangelo.makeart.turtlegenerator.maze.Generator_MazeCircle;
+import com.marginallyclever.makelangelo.makeart.turtlegenerator.maze.Generator_MazeRectangle;
+
 public class TurtleGeneratorFactory {
 	public static TurtleGenerator [] available = {
 			new Generator_Border(),
