+++ src/main/java/com/marginallyclever/makelangelo/makeart/turtlegenerator/maze/Generator_MazeCircle.java
@@ -1,6 +1,7 @@
-package com.marginallyclever.makelangelo.makeart.turtlegenerator;
+package com.marginallyclever.makelangelo.makeart.turtlegenerator.maze;
 
 import com.marginallyclever.makelangelo.Translator;
+import com.marginallyclever.makelangelo.makeart.turtlegenerator.TurtleGenerator;
 import com.marginallyclever.makelangelo.select.SelectReadOnlyText;
 import com.marginallyclever.makelangelo.select.SelectSlider;
 import com.marginallyclever.makelangelo.turtle.Turtle;
@@ -16,34 +17,9 @@ import java.util.Stack;
  * @author Dan Royer
  */
 public class Generator_MazeCircle extends TurtleGenerator {
-
-	// MazeWalls separate MazeCells
-	protected static class MazeWall {
-		int cellA, cellB;
-		boolean removed=false;
-
-		@Override
-		public String toString() {
-			return "["+cellA+","+cellB+"]";
-		}
-	}
-	// MazeCells are the rooms separted by MazeWalls
-	protected static class MazeCell {
-		int x, y;
-		boolean visited = false;
-
-		List<MazeWall> walls = new ArrayList<>();
-
-		@Override
-		public String toString() {
-			return "("+y+","+x+")"+(visited?"v":"");
-		}
-	}
-
-	protected static int rings = 5;
-	protected double xMax, xMin, yMax, yMin;
-	protected final List<MazeCell> cells = new ArrayList<>();
-	protected final List<MazeWall> walls = new ArrayList<>();
+	private static int rings = 5;
+	private final List<MazeCell> cells = new ArrayList<>();
+	private final List<MazeWall> walls = new ArrayList<>();
 
 	public Generator_MazeCircle() {
 		super();
@@ -210,10 +186,10 @@ public class Generator_MazeCircle extends TurtleGenerator {
 
 	private Turtle drawMaze() {
 		// find radius of maze
-		yMin = myPaper.getMarginBottom();
-		yMax = myPaper.getMarginTop();
-		xMin = myPaper.getMarginLeft();
-		xMax = myPaper.getMarginRight();
+		double yMin = myPaper.getMarginBottom();
+		double yMax = myPaper.getMarginTop();
+		double xMin = myPaper.getMarginLeft();
+		double xMax = myPaper.getMarginRight();
 
 		double r = Math.min(xMax-xMin,yMax-yMin)/2.0;
