+++ src/main/java/com/marginallyclever/makelangelo/makeart/turtlegenerator/maze/MazeWall.java
@@ -1,0 +1,16 @@
+package com.marginallyclever.makelangelo.makeart.turtlegenerator.maze;
+
+/**
+ * {@link MazeWall}s separate {@link MazeCell}s.
+ * @since 7.43.0
+ * @author Dan Royer
+  */
+class MazeWall {
+    public int cellA, cellB;
+    public boolean removed = false;
+
+    @Override
+    public String toString() {
+        return "[" + cellA + "," + cellB + "]";
+    }
+}
